Refactor code out to generate_return_code().
commit-id:40743e30
